+++ Changelog.md
@@ -1,4 +1,7 @@
-Version numbers are synced between hello_imgui and imgui_bundle.
+*Version numbers are synced between "Dear ImGui", "Hello ImGui" and "Dear ImGui Bundle"*
+
+# Unreleased
+* Separate ImageFromAsset and ImageFromAssetWithBg: use ImageFromAssetWithBg to display images with a background or border. 
 
 # v1.92.0
 Version numbers are now synced between "Dear ImGui" "Hello ImGui" and "Dear ImGui Bundle".